Create shared buffer to producers and consumer
This commit create a shared buffer to allow producers to send data to be
saved in SD as blocks.

The structure implemented allows the producer to notify the consumer
when data is available through semaphores, besides sharing the file
and the data (pointer and size in blocks) to be written.

Now the other acquisition channels can be implemented following the
same procedure of data sharing examplified in logger_analog_ch.c's
function io_adc_conv_callback.